+++ gvm/protocols/gmpv214/types.py
@@ -26,7 +26,6 @@ from gvm.protocols.gmpv208.types import (
     FeedType,
     FilterType,
     HostsOrdering,
-    PermissionSubjectType,
     SortOrder,
     TicketStatus,
     TimeUnit,
@@ -34,7 +33,6 @@ from gvm.protocols.gmpv208.types import (
     get_feed_type_from_string,
     get_filter_type_from_string,
     get_hosts_ordering_from_string,
-    get_permission_subject_type_from_string,
     get_sort_order_from_string,
     get_ticket_status_from_string,
     get_time_unit_from_string,
@@ -46,7 +44,6 @@ __all__ = [
     "FeedType",
     "FilterType",
     "HostsOrdering",
-    "PermissionSubjectType",
     "SeverityLevel",
     "SortOrder",
     "TicketStatus",
@@ -55,7 +52,6 @@ __all__ = [
     "get_feed_type_from_string",
     "get_filter_type_from_string",
     "get_hosts_ordering_from_string",
-    "get_permission_subject_type_from_string",
     "get_severity_level_from_string",
     "get_sort_order_from_string",
     "get_ticket_status_from_string",
